Refactor out base test class of linked transactions
